Fix bug in Item Activity and fix warning of Long to Integer in several methods

+++ nps/src/main/java/com/dpcat237/nps/dialog/PlayerLabelsDialog.java
@@ -42,7 +42,7 @@ public class PlayerLabelsDialog extends Activity {
         LabelRepository labelRepo = new LabelRepository(this);
         labelRepo.open();
 
-        Long itemId = intent.getLongExtra(ItemConstants.ITEM_ID, 0);
+        Integer itemId = intent.getIntExtra(ItemConstants.ITEM_ID, 0);
         item = itemRepo.getItem(itemId);
 
         ListView listView = (ListView) findViewById(R.id.labelsList);
